feat : 배송 서비스 dto 및 서비스 수정 (#6)

+++ src/main/java/org/pokeherb/deliveryservice/application/command/DeliveryCommandService.java
@@ -17,7 +17,7 @@ public interface DeliveryCommandService {
 
         void deleteDelivery(UUID deliveryId, String username);
 
-        void completeDelivery(UUID deliveryId, Integer actualDurationMin, Integer actualDurationKm);
+        void completeDelivery(UUID deliveryId, Double actualDurationMin, Double actualDurationKm);
 
         void changeRouteStatus(UUID deliveryId, Long routeId, DeliveryRouteStatusUpdateRequestDto requestDto);
 }
